Move the "inject-metadata" function to generic test
For pbench tests we are injecting our metadata to each result for later
analysis. Extracting this function and putting it to a BaseTest should
allow all tests that generate pbench-like results to use this function.
It's particularly important as we intend to add more metadata like
versions in the future.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>