Collect and present the system and profile environment
In case of a change it's really useful to know what all changed. This
commit brings support for "world" and "profile" information where
"world" environment means the host os as well as host runperf
configuration. The "profile" one is similar only it adds variout
profile-related tweaks that might have changed between executions.

The machine sysinfo collection is extendable via
"runperf.machine.distro_info" entry-point, the profiles collection is
only extensible via custom profiles.

Note the collection happens at the end of the execution as we usually
expect new packages to be installed by tests. In theory this might
result in multiple versions being lost between tests execution but it's
unlikely and the per-test collectibles are to be added.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>