Add fixtures for dynamic test-params
Certain test params have a dynamic nature and are expected to change,
which is making the comparison less-useful. Let's add fixtures to the
html-results report to only include the number of "clients" instead of
the full "clients" list. Note the raw values are not affected so users
can still get these via click.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>