utils: Use 1m load for wait_for_machine_calms_down
The loadavg is a moving average and calming down from load about 20 to 0
can take tenths of minuts while we are looking more for something in a
matter of minutes. Let's use the 1m load where calming down from load 40
to 0 takes about 4m, which should be plenty of time for the machine to
calm down.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>