tests: Fix the linpack test implementation
The linpack test implementation never really worked as pbench-linpack is
broken. It is possible to execute linpack via pbench-run-benchmark,
though, so let's use it.

There is currently one pbench bug which requires to execute:

    . /opt/pbench-agent/base

first so let's do that all the times as it should not harm anything.
Also at least on some systems (RHEL8) the default python for ansible is
still python2 but using it would require special handling so let's
hardcode the ANSIBLE_PYTHON_INTERPRETER to python3 for now as well as
disable ANSIBLE_HOST_KEY_CHECKING to avoid the need to exchange keys
with worker machines. When everything settles down we'll hopefully be
able to remove these.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>