Use git describe-based versioning scheme
Use "git describe" to either get the current tag or the closest tag with
hash. On top of that also check for dirtiness and append the "-dirty" in
such case.

As a result we should get versions like:

    0.9
    0.9-dirty
    0.9-95-g5c6fd51-dirty

which should conform with the usual python versioning while allowing
to parse the actual commit/tag from the output.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>