Allow to store the raw "get_info" values along with the filtered ones
The "get_info" has two main usages, to store the actual configuration
and to be used for diff purposes. For diff purposes we need certain
fields filtered so the dynamic things like ip addresses won't result in
unnecessary diffs but we might still be interested in the actual raw
values in order to reproduce the configuration.

Let's store the raw values with "$key_raw" prefix along with the
filtered value being stored just under the $key.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>