html_result: Report all machines/environment entries
Historically we only reported the first machine as well as profile
environments. Let's turn these into lists and report the details per
each item.

For machines this means reporting links for each machine, for profiles
this means reporting short name and diff per each machine and allow to
copy the full environment (as it is usually better to get list of all
envs than just a single env raw values).

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>