html_result: Sort the env items
this should make sure the reported items are in the same order so we can
more rely on positioning rather than checking each item individually.

Note when the number of params differs or when there are missing keys we
will end up with things on a different place but it should be stable
enough for a quick glance.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>