Add linear regression model based on standard deviation
The current linear regression model uses minimum and maximum values
present in the sample. This requires careful picking of the sample and
is prone to outliers. Let's use 3x standard deviation which is the usual
definition of an outlier. Note that by using different "tolerance" you
can increase or decrease the tolerance after the model is applied.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>