Allow multiple models to add to the overall test score
previously we only used either the first model value or raw value as the
test score. Let's allow for models to report not just score but also a
weight of this particular score and calculate the overall test score
using those weights.

At this point the system is not much flexible and specifies a static
weights, but in the future we do want to let users and models to specify
their custom scores.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>