Add all-builds average check
´In order to spot mutliple-builds smaller improvements/regressions (eg.
if >1/2 of builds is just under the treshold) as well as to smooth
single build outliers (in case the average value is +- OK but this build
performed slightly worse) let's add an all-builds average check.

As this works on the relative score and not the absolute values it
couldn't be implemented as a model, therefor it's an additional check
inside the RelativeResults.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>