html_report: Process the environment diff from the dst result
As the src build is usually a baseline, there can be quite a lot of
changes between it and the individual results. On the other hand when
ran regularly the diff between reference builds and the dst build is
usually smaller. Let's use the dst result as a source for the env diffs.

In order to emphasize the order we need to reverse the processing order
so the indexes start from back.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>