diffperf: Use stddev-based probability of samples
For tests where stddev is available use the probability of the each
result to calculate the similarity score.

This should provide better results in the case where some tests perform
alike as these should get equal score unlike when we were normalizing
the distances.

On the other hand this will ignore the results of tests that used a
single sample, so make sure to always compare all tests with stddevs or
no stddevs at all, otherwise you might get results based on the few
tests with stddevs ignoring the others.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>