Add our profile to the processed test_params
In order to distinguish between potentially different test params we are
using the pbench's "benchmark.parameters[]". Let's add the
"benchmark.user[]" processing to double-check we are using the profile
we thing we are using.

It happened that different results were re-used and in such case this
could improve the detection.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>