Add support to fetch_logs and use it on profile failures
In case of a apply/revert_profile failure there might be essential
files/commands that should help identifying the issues. Let's allow to
define them in the process handler and raise StepFailed in such
occassion. To visualize the failures let's store them in a defined
structure and process this dir structure in results comparison.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>