closest_result: Fix the closest result match
We are using a selection to only find the closest match on the
sub-selection of results based on previous level, but when reporting the
value we were simply searching for the first match. In case there was a
match that was not in the selection, we still returned, which resulted
in odd results.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>