closest_result: Add a note about range of tests used for comparison
the usage of missing tests only increases the likelihood of spoiling the
results (tested and then well thought). We want to find the closest
result to our current src result based **only** on the src test results
and ignore the extra tests that might have been executed on the restults
we are comparing to.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>