Add "strip-run-perf" script
This script can be used to create a stripped version of the run-perf
results. It removes all the information which is not used by
compare-perf like individual sample values or other assets, but for
compare-perf (and other run-perf related scripts) it should preserve all
of the values. The primary focus is to allow compare-perf/diff-perf/...
execution on a minimal results when the full pbench data are stored
somewhere else.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>