docs.jenkins: Execute prune-artifacts daily rather than on runperf end
The prune-artifacts job is used to cleanup results after a certain
period of time. Running it after job execution works well for daily jobs
but doesn't work for ad-hoc cases (as it's not executed regularly,
therefore after 14 days it won't prune the results).

Let's run the prune job daily and specify all the jobs manually.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>