docs.jenkins: Increase the number of referenceBuilds
As the runperf pipeline now includes the current build to the reference
builds we need to increase the number of reference builds.

We could also just remove the current build from runperf reference
builds but including it allows easily check which failures are related
to averages and which to model results.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>