Use uncertainty for group results tolerances
The group results combine multiple results to represent certain group.
The more results in the same group there are the greater probability of
the spread to be closer to 0. Let's use the uncertainty we use elsewhere
to decrease the tolerance.

As we want to be stricter than with normal results divide the resulting
tolerance by 2 (for now, maybe consider customization if needed).

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>