Add pbench cleanup on failure
pbench-0.71 on error usually leaves the redis and tool-meister running,
which collides with the newly created instances in the following tests.
As pbench team is not sure when they will fix this issue, let's address
this by nuking potential processes in such case.

Pbench issue: https://github.com/distributed-system-analysis/pbench/
              issues/2625

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>