bisect: Add 2 out of 3 mode
In case the output might occasionally wander off we might want to use
the 2 out of 3 mode for checking whether a certain build is good or bad.
This prolongs the execution due to the need to execute the same build
twice or even three-times in case of a different first two results but
it significantly decreases the chance of going the wrong direction by
accidental result.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>