Add support to rebase a model
In case of a performance change we sometimes just need to quickly move
the baseline while keeping the acceptable deviation. This can be
achieved by using "--rebase-model" which uses the original model and
changes only it's raw values. Such model should only be used for a
transient period but, provided the jitter stays similar, should does the
job well.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>