Add support for group of diff-perf results
Groups of results are useful when we have jittery results and want to
average across multiple collected results. The way it works is we
bundle/group multiple results into the same group and diff-perf
calculates the usual distance (per test). Before getting the closest
distance we average each group distances:

    diff-perf src dst1 dst2 -g dst3 dst4 -g dst5 dst6 dst7

tries to match src to:

0) dst
1) dst2
2) average of dst3, dst4
3) average of dst5, dst6, dst7

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>