docs.jenkins: Decrease diff-perf verbosity
We don't usually need the full run-perf outputs from the individual
runs, let's store them only in the workspace and use only "-v" in the
jenkins console output.

Note sometimes the log files might be useful, in that case one can check
the last-job's one in the job workspace, but we don't archive them as in
most cases they are useless (similarly to the job results). One can
re-run these manually if needed.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>