Allow pbench tests to fail early on no new output
often our tests consist of multiple iterations of a well defined
duration. Let's add a watchdog for a new output produced by the executed
test and fail early when we don't get any update in the expected time.
Use a minimum of 60s timeout to avoid failing on sysinfo collection.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>