docs.jenkins: Fix qemu deployment in bisect pipeline
In bisection pipeline we have to deploy qemu prior to bisection
execution in order to be able to run the same profile in the pre-job.
Afterwards we have to supply the original setup script without qemu for
the actual bisection job as the bisection script adds the qemu
deployment itself.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>